backend/API/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-jose==3.3.0
pydantic==2.10.3
email_validator==2.2.0